+++ pgx/src/trigger_support/mod.rs
@@ -1,0 +1,83 @@
+/*
+Portions Copyright 2019-2021 ZomboDB, LLC.
+Portions Copyright 2021-2022 Technology Concepts & Design, Inc. <support@tcdi.com>
+
+All rights reserved.
+
+Use of this source code is governed by the MIT license that can be found in the LICENSE file.
+*/
+
+//! Helper functions for working with custom Rust trigger functions
+
+mod pg_trigger;
+mod pg_trigger_error;
+mod pg_trigger_level;
+mod pg_trigger_option;
+mod pg_trigger_safe;
+mod pg_trigger_when;
+mod trigger_tuple;
+
+pub use pg_trigger::PgTrigger;
+pub use pg_trigger_error::PgTriggerError;
+pub use pg_trigger_level::PgTriggerLevel;
+pub use pg_trigger_option::PgTriggerOperation;
+pub use pg_trigger_safe::PgTriggerSafe;
+pub use pg_trigger_when::PgTriggerWhen;
+pub use trigger_tuple::TriggerTuple;
+
+use crate::{is_a, pg_sys};
+
+/// A newtype'd wrapper around a `pg_sys::TriggerData.tg_event` to prevent accidental misuse
+#[derive(Debug)]
+pub struct TriggerEvent(u32);
+
+#[inline]
+pub unsafe fn called_as_trigger(fcinfo: pg_sys::FunctionCallInfo) -> bool {
+    let fcinfo = fcinfo.as_ref().expect("fcinfo was null");
+    !fcinfo.context.is_null() && is_a(fcinfo.context, pg_sys::NodeTag_T_TriggerData)
+}
+
+#[inline]
+pub fn trigger_fired_by_insert(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_INSERT
+}
+
+#[inline]
+pub fn trigger_fired_by_delete(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_DELETE
+}
+
+#[inline]
+pub fn trigger_fired_by_update(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_UPDATE
+}
+
+#[inline]
+pub fn trigger_fired_by_truncate(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_TRUNCATE
+}
+
+#[inline]
+pub fn trigger_fired_for_row(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_ROW != 0
+}
+
+#[inline]
+pub fn trigger_fired_for_statement(event: u32) -> bool {
+    !trigger_fired_for_row(event)
+}
+
+#[inline]
+pub fn trigger_fired_before(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_TIMINGMASK == pg_sys::TRIGGER_EVENT_BEFORE
+}
+
+#[inline]
+pub fn trigger_fired_after(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_TIMINGMASK == pg_sys::TRIGGER_EVENT_AFTER
+}
+
+#[inline]
+pub fn trigger_fired_instead(event: u32) -> bool {
+    event & pg_sys::TRIGGER_EVENT_TIMINGMASK == pg_sys::TRIGGER_EVENT_INSTEAD
+}
