+++ pgx/src/trigger_support/pg_trigger.rs
@@ -1,22 +1,14 @@
-/*
-Portions Copyright 2019-2021 ZomboDB, LLC.
-Portions Copyright 2021-2022 Technology Concepts & Design, Inc. <support@tcdi.com>
-
-All rights reserved.
-
-Use of this source code is governed by the MIT license that can be found in the LICENSE file.
-*/
-
-//! Helper functions for working with custom Rust trigger functions
-
-use cstr_core::c_char;
-
 use crate::{
     heap_tuple::PgHeapTuple,
-    is_a, pg_sys,
+    pg_sys,
     pgbox::{AllocatedByPostgres, PgBox},
-    PgRelation,
+    rel::PgRelation,
+    trigger_support::{
+        called_as_trigger, PgTriggerError, PgTriggerLevel, PgTriggerOperation, PgTriggerSafe,
+        PgTriggerWhen, TriggerEvent, TriggerTuple,
+    },
 };
+use cstr_core::c_char;
 use std::borrow::Borrow;
 
 /**
@@ -404,232 +396,4 @@ impl PgTrigger {
 
         Ok(trigger_safe)
     }
-}
-
-pub struct PgTriggerSafe<'a> {
-    pub name: &'a str,
-    pub new: Option<PgHeapTuple<'a, AllocatedByPostgres>>,
-    pub current: Option<PgHeapTuple<'a, AllocatedByPostgres>>,
-    pub event: TriggerEvent,
-    pub when: PgTriggerWhen,
-    pub level: PgTriggerLevel,
-    pub op: PgTriggerOperation,
-    pub relid: pg_sys::Oid,
-    pub old_transition_table_name: Option<&'a str>,
-    pub new_transition_table_name: Option<&'a str>,
-    pub relation: crate::PgRelation,
-    pub table_name: String,
-    pub table_schema: String,
-    pub extra_args: Vec<String>,
-}
-
-/// A newtype'd wrapper around a `pg_sys::TriggerData.tg_event` to prevent accidental misuse
-#[derive(Debug)]
-pub struct TriggerEvent(u32);
-
-/// When a trigger happened
-///
-/// Maps from a `TEXT` of `BEFORE`, `AFTER`, or `INSTEAD OF`.
-///
-/// Can be calculated from a `pgx_pg_sys::TriggerEvent`.
-// Postgres constants: https://cs.github.com/postgres/postgres/blob/36d4efe779bfc7190ea1c1cf8deb0d945b726663/src/include/commands/trigger.h?q=TRIGGER_FIRED_BEFORE#L100-L102
-// Postgres defines: https://cs.github.com/postgres/postgres/blob/36d4efe779bfc7190ea1c1cf8deb0d945b726663/src/include/commands/trigger.h?q=TRIGGER_FIRED_BEFORE#L128-L135
-pub enum PgTriggerWhen {
-    /// `BEFORE`
-    Before,
-    /// `AFTER`
-    After,
-    /// `INSTEAD OF`
-    InsteadOf,
-}
-
-impl TryFrom<TriggerEvent> for PgTriggerWhen {
-    type Error = PgTriggerError;
-    fn try_from(event: TriggerEvent) -> Result<Self, Self::Error> {
-        match event.0 & pg_sys::TRIGGER_EVENT_TIMINGMASK {
-            pg_sys::TRIGGER_EVENT_BEFORE => Ok(Self::Before),
-            pg_sys::TRIGGER_EVENT_AFTER => Ok(Self::After),
-            pg_sys::TRIGGER_EVENT_INSTEAD => Ok(Self::InsteadOf),
-            v => Err(PgTriggerError::InvalidPgTriggerWhen(v)),
-        }
-    }
-}
-
-impl ToString for PgTriggerWhen {
-    fn to_string(&self) -> String {
-        match self {
-            PgTriggerWhen::Before => "BEFORE",
-            PgTriggerWhen::After => "AFTER",
-            PgTriggerWhen::InsteadOf => "INSTEAD OF",
-        }
-        .to_string()
-    }
-}
-
-/// The level of a trigger
-///
-/// Maps from a `TEXT` of `ROW` or `STATEMENT`.
-///
-/// Can be calculated from a `pgx_pg_sys::TriggerEvent`.
-// Postgres constants: https://cs.github.com/postgres/postgres/blob/36d4efe779bfc7190ea1c1cf8deb0d945b726663/src/include/commands/trigger.h?q=TRIGGER_FIRED_BEFORE#L98
-// Postgres defines: https://cs.github.com/postgres/postgres/blob/36d4efe779bfc7190ea1c1cf8deb0d945b726663/src/include/commands/trigger.h?q=TRIGGER_FIRED_BEFORE#L122-L126
-pub enum PgTriggerLevel {
-    /// `ROW`
-    Row,
-    /// `STATEMENT`
-    Statement,
-}
-
-impl From<TriggerEvent> for PgTriggerLevel {
-    fn from(event: TriggerEvent) -> Self {
-        match event.0 & pg_sys::TRIGGER_EVENT_ROW {
-            0 => PgTriggerLevel::Statement,
-            _ => PgTriggerLevel::Row,
-        }
-    }
-}
-
-impl ToString for PgTriggerLevel {
-    fn to_string(&self) -> String {
-        match self {
-            PgTriggerLevel::Statement => "STATEMENT",
-            PgTriggerLevel::Row => "ROW",
-        }
-        .to_string()
-    }
-}
-
-/// The operation for which the trigger was fired
-///
-/// Maps from a `TEXT` of `INSERT`, `UPDATE`, `DELETE`, or `TRUNCATE`.
-///
-/// Can be calculated from a `pgx_pg_sys::TriggerEvent`.
-// Postgres constants: https://cs.github.com/postgres/postgres/blob/36d4efe779bfc7190ea1c1cf8deb0d945b726663/src/include/commands/trigger.h#L92
-// Postgres defines: https://cs.github.com/postgres/postgres/blob/36d4efe779bfc7190ea1c1cf8deb0d945b726663/src/include/commands/trigger.h#L92
-pub enum PgTriggerOperation {
-    /// `INSERT`
-    Insert,
-    /// `UPDATE`
-    Update,
-    /// `DELETE`
-    Delete,
-    /// `TRUNCATE`
-    Truncate,
-}
-
-impl TryFrom<TriggerEvent> for PgTriggerOperation {
-    type Error = PgTriggerError;
-    fn try_from(event: TriggerEvent) -> Result<Self, Self::Error> {
-        match event.0 & pg_sys::TRIGGER_EVENT_OPMASK {
-            pg_sys::TRIGGER_EVENT_INSERT => Ok(Self::Insert),
-            pg_sys::TRIGGER_EVENT_DELETE => Ok(Self::Delete),
-            pg_sys::TRIGGER_EVENT_UPDATE => Ok(Self::Update),
-            pg_sys::TRIGGER_EVENT_TRUNCATE => Ok(Self::Truncate),
-            v => Err(PgTriggerError::InvalidPgTriggerOperation(v)),
-        }
-    }
-}
-
-impl ToString for PgTriggerOperation {
-    fn to_string(&self) -> String {
-        match self {
-            PgTriggerOperation::Insert => "INSERT",
-            PgTriggerOperation::Update => "UPDATE",
-            PgTriggerOperation::Delete => "DELETE",
-            PgTriggerOperation::Truncate => "TRUNCATE",
-        }
-        .to_string()
-    }
-}
-
-#[derive(thiserror::Error, Debug, Clone, Copy)]
-pub enum PgTriggerError {
-    #[error("`PgTrigger`s can only be built from `FunctionCallInfo` instances which `pgx::pg_sys::called_as_trigger(fcinfo)` returns `true`")]
-    NotTrigger,
-    #[error("`PgTrigger`s cannot be built from `NULL` `pgx::pg_sys::FunctionCallInfo`s")]
-    NullFunctionCallInfo,
-    #[error(
-        "`InvalidPgTriggerWhen` cannot be built from `event & TRIGGER_EVENT_TIMINGMASK` of `{0}"
-    )]
-    InvalidPgTriggerWhen(u32),
-    #[error(
-        "`InvalidPgTriggerOperation` cannot be built from `event & TRIGGER_EVENT_OPMASK` of `{0}"
-    )]
-    InvalidPgTriggerOperation(u32),
-    #[error("core::str::Utf8Error: {0}")]
-    CoreUtf8(#[from] core::str::Utf8Error),
-    #[error("TryFromIntError: {0}")]
-    TryFromInt(#[from] core::num::TryFromIntError),
-    #[error("The `pgx::pg_sys::TriggerData`'s `tg_trigger` field was a NULL pointer")]
-    NullTrigger,
-    #[error("The `pgx::pg_sys::FunctionCallInfo`'s `context` field was a NULL pointer")]
-    NullTriggerData,
-    #[error("The `pgx::pg_sys::TriggerData`'s `tg_relation` field was a NULL pointer")]
-    NullRelation,
-}
-
-/// Indicates which trigger tuple to convert into a [crate::PgHeapTuple].
-#[derive(Debug, Copy, Clone)]
-pub enum TriggerTuple {
-    /// The row for which the trigger was fired. This is the row being inserted, updated, or deleted.
-    /// If this trigger was fired for an INSERT or DELETE then this is what you should return from the
-    /// function if you don't want to replace the row with a different one (in the case of INSERT) or
-    /// skip the operation.
-    Current,
-
-    /// The new version of the row, if the trigger was fired for an UPDATE. This is what you have to
-    /// return from the function if the event is an UPDATE and you don't want to replace this row by
-    /// a different one or skip the operation.
-    New,
-}
-
-#[inline]
-pub unsafe fn called_as_trigger(fcinfo: pg_sys::FunctionCallInfo) -> bool {
-    let fcinfo = fcinfo.as_ref().expect("fcinfo was null");
-    !fcinfo.context.is_null() && is_a(fcinfo.context, pg_sys::NodeTag_T_TriggerData)
-}
-
-#[inline]
-pub fn trigger_fired_by_insert(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_INSERT
-}
-
-#[inline]
-pub fn trigger_fired_by_delete(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_DELETE
-}
-
-#[inline]
-pub fn trigger_fired_by_update(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_UPDATE
-}
-
-#[inline]
-pub fn trigger_fired_by_truncate(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_OPMASK == pg_sys::TRIGGER_EVENT_TRUNCATE
-}
-
-#[inline]
-pub fn trigger_fired_for_row(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_ROW != 0
-}
-
-#[inline]
-pub fn trigger_fired_for_statement(event: u32) -> bool {
-    !trigger_fired_for_row(event)
-}
-
-#[inline]
-pub fn trigger_fired_before(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_TIMINGMASK == pg_sys::TRIGGER_EVENT_BEFORE
-}
-
-#[inline]
-pub fn trigger_fired_after(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_TIMINGMASK == pg_sys::TRIGGER_EVENT_AFTER
-}
-
-#[inline]
-pub fn trigger_fired_instead(event: u32) -> bool {
-    event & pg_sys::TRIGGER_EVENT_TIMINGMASK == pg_sys::TRIGGER_EVENT_INSTEAD
 }
